ref: move release fee to factory from router
