+++ contracts/liquidity/factory/src/state.rs
@@ -1,5 +1,5 @@
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{Addr, Binary, Int256, Uint128};
+use cosmwasm_std::{Addr, Binary, Decimal, Int256, Uint128};
 use cw_storage_plus::{Item, Map};
 use euclid::{
     chain::ChainUid,
@@ -84,3 +84,6 @@ pub const COSMOS_PACKET_RELAY_SEQUENCE_COUNT: Item<u128> =
     Item::new("cosmos_packet_relay_sequence_count");
 
 pub const PROCESSED_PACKET_SEQUENCE: Map<u128, Uint128> = Map::new("processed_packet_sequence");
+
+/// The key is TokenID_ChainUID
+pub const RELEASE_FEES: Map<String, Decimal> = Map::new("release_fees");
